+++ client-app/src/desktop/common/grid/SampleGridModel.ts
@@ -200,7 +200,7 @@ export class SampleGridModel extends HoistModel {
     }
 
     override async doLoadAsync(loadSpec) {
-        const {trades} = await XH.fetchJson({url: 'trade', correlationId: true}),
+        const {trades} = await XH.fetchJson({url: 'trade'}),
             {gridModel} = this;
 
         gridModel.loadData(trades);