+++ client-app/src/desktop/tabs/grids/ColumnFilteringPanelModel.ts
@@ -22,7 +22,7 @@ export class ColumnFilteringPanelModel extends HoistModel {
     }
 
     override async doLoadAsync(loadSpec) {
-        const {trades} = await XH.fetchJson({url: 'trade', correlationId: true}),
+        const {trades} = await XH.fetchJson({url: 'trade'}),
             {gridModel} = this;
 
         gridModel.loadData(trades);