+++ client-app/src/desktop/tabs/other/PlaceholderPanel.tsx
@@ -86,7 +86,7 @@ class PlaceholderPanelModel extends HoistModel {
     });
 
     override async doLoadAsync(loadSpec) {
-        const {trades} = await XH.fetchJson({url: 'trade', correlationId: true});
+        const {trades} = await XH.fetchJson({url: 'trade'});
         this.gridModel.loadData(trades);
     }
 }