+++ client-app/src/examples/news/NewsPanelModel.ts
@@ -52,7 +52,7 @@ export class NewsPanelModel extends HoistModel {
     }
 
     override async doLoadAsync(loadSpec) {
-        const stories = await XH.fetchJson({url: 'news', loadSpec, correlationId: true});
+        const stories = await XH.fetchJson({url: 'news', loadSpec});
         this.completeLoad(stories);
     }