+++ client-app/src/mobile/grids/GridDetailPage.ts
@@ -53,8 +53,6 @@ class GridDetailPageModel extends HoistModel {
     }
 
     override async doLoadAsync() {
-        await XH.fetchJson({url: 'customer', correlationId: true}).thenAction(
-            c => (this.customers = c)
-        );
+        await XH.fetchJson({url: 'customer'}).thenAction(c => (this.customers = c));
     }
 }